# api/openapi.yaml
openapi: 3.0.3
info:
  title: Writers' Guild API
  version: 0.1.0
  description: |
    HTTP contract of the Writers' Guild. The Go server interface and the
    TypeScript client are generated from this file. Streaming endpoints use
    Server-Sent Events and are documented here but implemented by hand.
servers:
  - url: /
security: []
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
  /api/gateway/models:
    get:
      operationId: listGatewayModels
      summary: Aliases known to the gateway
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayModels" }
        default: { $ref: "#/components/responses/Error" }
  /api/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectSummary" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
  /api/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      operationId: getProject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteProject
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
...
  /api/writers:
    get:
      operationId: listWriters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Writer" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createWriter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriterInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Writer" }
        default: { $ref: "#/components/responses/Error" }
  /api/writers/test:
    post:
      operationId: testWriter
      summary: Send a short sample request through the gateway
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriterTestInput" }
      responses:
        "200":
          description: The gateway answered or refused; see ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriterTestResult" }
        default: { $ref: "#/components/responses/Error" }
  /api/writers/{writerId}:
    parameters:
      - $ref: "#/components/parameters/writerId"
    get:
      operationId: getWriter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Writer" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateWriter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriterInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Writer" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteWriter
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
  /api/writers/{writerId}/duplicate:
    parameters:
      - $ref: "#/components/parameters/writerId"
    post:
      operationId: duplicateWriter
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Writer" }
        default: { $ref: "#/components/responses/Error" }
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    writerId:
      name: writerId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    Settings:
      type: object
      required: [scene_token_limit, autosave_snapshot_minutes]
      properties:
        scene_token_limit: { type: integer }
        autosave_snapshot_minutes: { type: integer }
    SettingsInput:
      type: object
      required: [scene_token_limit, autosave_snapshot_minutes]
      properties:
        scene_token_limit: { type: integer, minimum: 500, maximum: 200000 }
        autosave_snapshot_minutes: { type: integer, minimum: 1, maximum: 1440 }
    GatewayModels:
      type: object
      required: [alias_prefix, aliases, all_models]
      properties:
        alias_prefix: { type: string }
        aliases:
          type: array
          description: Models whose id starts with the alias prefix
          items: { type: string }
        all_models:
          type: array
          items: { type: string }
    Project:
      type: object
      required: [id, name, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectSummary:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required: [chapter_count]
          properties:
            chapter_count: { type: integer }
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000 }
...
    WriterRole:
      type: string
      enum: [critic, co-writer]
    Writer:
      type: object
      required: [id, name, slug, model_alias, system_prompt, roles, enabled, temperature, is_system, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        model_alias: { type: string }
        system_prompt: { type: string }
        roles:
          type: array
          items: { $ref: "#/components/schemas/WriterRole" }
        enabled: { type: boolean }
        temperature: { type: number, format: double }
        is_system: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WriterInput:
      type: object
      required: [name, system_prompt, roles, enabled, temperature]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        model_alias:
          type: string
          description: Defaults to "{APP_NAME}-{slug}" when empty on create
        system_prompt: { type: string, maxLength: 20000 }
        roles:
          type: array
          items: { $ref: "#/components/schemas/WriterRole" }
        enabled: { type: boolean }
        temperature: { type: number, format: double, minimum: 0, maximum: 2 }
    WriterTestInput:
      type: object
      required: [name, model_alias, system_prompt, temperature]
      properties:
        writer_id: { type: string, format: uuid }
        name: { type: string }
        model_alias: { type: string, minLength: 1 }
        system_prompt: { type: string }
        temperature: { type: number, format: double, minimum: 0, maximum: 2 }
    WriterTestResult:
      type: object
      required: [ok, run_id, model_alias, prompt_tokens, completion_tokens, cost_known, cost_estimated, latency_ms]
      properties:
        ok: { type: boolean }
        run_id: { type: string, format: uuid }
        model_alias: { type: string }
        reply: { type: string }
        error: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        cost_usd: { type: number, format: double }
        cost_known: { type: boolean }
        cost_estimated: { type: boolean }
        latency_ms: { type: integer }